refactor(InputBox): merge useInputKeyboardHandler into InputBox component
